Change adder contract to be u64 values

+++ examples/adder.sol
@@ -2,7 +2,7 @@ pragma solidity 0;
 
 contract adder {
 	/// Add the values of x and y together.
-	function add(uint32 x, uint32 y) public pure returns (uint32) {
+	function add(uint64 x, uint64 y) public pure returns (uint64) {
 		return x + y;
 	}
 }
